Move download logic into DownloadButton component
