+++ src/components/alphabet-view/selected-letter-display.tsx
@@ -1,17 +1,20 @@
 import React from 'react';
+import { DictionaryEntryExportFormat } from '@mothertongues/search';
+
+// FPCC
 import { CopyButton } from './copy-button';
 import { DownloadButton } from './download-button';
 import { AudioButton } from './audio-button';
-import { FvLetter, FVMedia } from '../common/data';
+import { FvLetter, FVMedia, FvWordLocationCombo } from '../common/data';
 
 export interface SelectedLetterDisplayProps {
-  selected: FvLetter | null;
-  promptForDownload: () => void;
+  dataDictionary: (DictionaryEntryExportFormat | FvWordLocationCombo)[];
+  selected: FvLetter;
 }
 
 export function SelectedLetterDisplay({
   selected,
-  promptForDownload,
+  dataDictionary,
 }: Readonly<SelectedLetterDisplayProps>) {
   const audioCount = selected?.relatedAudio.length ?? 0;
 
@@ -21,13 +24,13 @@ export function SelectedLetterDisplay({
       {audioCount === 0 && (
         <div className="flex w-full justify-center">
           <CopyButton selected={selected} />
-          <DownloadButton promptForDownload={promptForDownload} />
+          <DownloadButton selected={selected} dataDictionary={dataDictionary} />
         </div>
       )}
       {audioCount === 1 && (
         <div className="grid grid-cols-3">
           <CopyButton selected={selected} />
-          <DownloadButton promptForDownload={promptForDownload} />
+          <DownloadButton selected={selected} dataDictionary={dataDictionary} />
           {selected?.relatedAudio.map((fvAudio: FVMedia) => {
             return <AudioButton key={fvAudio.id} fvAudio={fvAudio} />;
           })}
@@ -39,7 +42,10 @@ export function SelectedLetterDisplay({
             <CopyButton selected={selected} />
           </div>
           <div className="flex w-full justify-center">
-            <DownloadButton promptForDownload={promptForDownload} />
+            <DownloadButton
+              selected={selected}
+              dataDictionary={dataDictionary}
+            />
           </div>
           <div className="flex justify-evenly mt-5">
             {selected?.relatedAudio.map((fvAudio: FVMedia) => {
